Refactor restart of other OpenStack services
We restart some OpenStack services to allow a full OpenStack
installation up to and including a VNF activation through a single
openstack-chef-repo kitchen build.

With this patch, we add a guard against restarting these services
outside of the kitchen environment.

We call systemctl directly, which is a more straight-forward method of
achieving an immediate restart than going through Chef notification.