Create and use an OpenStack NFV orchestration demo user
Create a separate user (nfvdemo) to execute the sample scripts
(vim-sample, vnfd-sample, vnf-sample). This makes it theoretically
possible to reduce the rights granted to this demo user (a project admin
role, however, is needed to use OS::Nova::Flavor when creating the vnf).

The actual reason that triggered this change was openstack_command_env
(openstack-common cookbook). It looks for the user password in the
user_passwords vault which is not where the tacker (service) user's
password is stored.

Alternative workarounds considered include a) a copy containing the
tacker service secret in the user_passwords vault and b) a replacement
function for openstack_command_env.